requirements.txt
numpy
langchain
langchain-community
langchain-huggingface
langchain-google-genai
transformers 
sentence-transformers
pypdf
tiktoken
fastapi[standard]
uvicorn
faiss-cpu